Anonymous vs Named Generic Implementation Parameter
In the chapter 8.1 "Generic Functions", anonymous generic implementation parameters
are described at the end of the chapter (in "Anonymous Generic Implementation Parameter (+ operator)"
but they are used in the second and third code snippets, at the beginning of the chapter.

Then, the chapter "Constraints for Generic Types" mentions a `TDrop` implementation but in the code
snippet, the drop implementation is anonymous.

To be more consistent with the chapter structure, this PR uses named generic implementation parameters
in the first code snippets, until anonymous generic implementation parameter are described.