Issue in wallet field types
At the end of the chapter 8.1, 2 `Wallet` are instanciated with the following code:

```
    let w1 = Wallet { balance: true, address: 10 };
    let w2 = Wallet { balance: 32, address: 100 };
```

Then, a paragraph explains that we create 2 instances: one of `Wallet<bool, u128>` and the other of `Wallet<felt252, u8>`.

I think that it's not correct as the compiler will use `felt252` to store `10`, `32` and `100` values by default.

To be sure to have an instance of `Wallet<bool, u128>` and another of `Wallet<felt252, u8>`, we have to use `10_u128` and `100_u8`.