Match catch-all operator
The chapter about `match` catch-all operator is not so clear as it talks about
`enums` and `felt252` while it is only supported for `felt252` in Cairo 2.4.

As `match` will be strongly improved in Cairo 2.5 according to the roadmap,
I think it would be better to just mention how it works for `felt252` for the moment,
and then deeply update this chapter for Cairo 2.5.

As I'm a Cairo newbie, I may have missed some information about `match` and so this PR
may be not fully correct.