Update ch02-01-variables-and-mutability.md
by default cairo's variables  are immutable and remain immutable, what happens under the hood using 'mut' is that, the corresponding Sierra variable storing its value is first dropped as it's no longer used, and then a new variable is created with the updated value. so we can say the mut keyword is used to make variables "act" in a mutable way. rather than saying "make them mutable"